MainMenuUpdate
Fixed Volume not saving issue by a genius way, and added controls (cant change controls), and a little tweaks here and there of random stuff and names. SOMETHING IS HUGELY WRONG when you hit Play on the main menu I have it going to the main map, but there is a really weird issue that happens. Lighting is off, character is super fast, things are very scuffed and idk how to fix that at all...