Inventory, Map, PauseMenu, CardBook Update
Made it so they all do not overlap each other. I like how the inventory can be closed using escape, that is not the case for the map or the card book and I would like to look into that but it is not necessary. Just have to use M or B to open and close the map and card book. BUT, THEY DO NOT OVERLAP AND CAN NOT BE USED SIMALTANEOUSLY ANYMORE WHICH IS REALLY NICE